// api/models/ImageProcessorResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ImageProcessorResponse = {
    /**
     * Base64-encoded PNG image produced by the processor.
     */
    image: string;
    /**
     * Processing duration in milliseconds.
     */
    duration_ms: number;
    /**
     * Model identifier returned by the processor.
     */
    model: string;
};